Estimate return period wind speeds from TCs
This script is largely the same as estimate_return_periods.py, with some
modifications to work with the observed gust wind speeds that are
attributable to tropical cyclone events. Main changes are the format of
the data files and the addition of an optional kwarg to set the minimum
threshold for selecting the threshold value.